Update schema proto file
This commit updates the schema protocol to use the latest version
present on [mafintosh/simple-hypercore-protocol](https://github.com/mafintosh/simple-hypercore-protocol/blob/d577298a47479630d0e340190273db4a22d3c793/schema.proto).

This includes the NOISE protocol handshake as part of the file, and it
is not backward compatible with DAT <= 7. Due to that, a major bump has
been proposed.

It uses a simplified single file/single namespace proposal, to help to
keep it up to date instead of checking multiple files to see if the
fields are updated, which could be split if needed.